requirements.txt
alembic==1.4.2
amqp-worker==0.2
axial-positional-embedding==0.2.1
certifi==2020.4.5.1
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
diff-match-patch==20120106
dill==0.3.3
docutils==0.16
future==0.18.2
h5py==2.10.0
imageio==2.8.0
imgaug==0.4.0
joblib==0.15.1
kiwisolver==1.2.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl-service==2.3.0
multiprocess==0.70.11.1
networkx==2.4
numpy==1.17.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
p-tqdm==1.3.3
pandas==1.0.4
pathos==0.2.7
pika==1.1.0
Pillow==7.1.2
pox==0.2.9
ppft==1.6.6.3
product-key-memory==0.1.10
progressbar==2.5
progressbar2==3.53.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.17
submitit==1.0.0
tabulate==0.8.7
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.6.3
tqdm==4.46.1
typing-extensions==3.7.4.2
ujson==3.0.0
